fixed issue #34 (sorting_package) (#36)
Refactor sorting
