+++ Sorting/MergeSort.java
@@ -1,53 +1,61 @@
-/*
- * Developed by Atri Tripathi on 14/7/19 9:39 PM
- * Last modified 14/7/19 9:39 PM
- * Copyright (c) 2019. All rights reserved
- */
+import java.util.Arrays;
 
-public class MergeSort {
-    private void sort(int[] arr, int lb, int ub) {
-        if (lb < ub) {
-            int mid = (lb + ub)/2;
-            sort(arr,lb,mid);
-            sort(arr,mid+1,ub);
+public class MergeSort<T extends Comparable<T>> {
 
-            merge(arr,lb,mid,ub);
-        }
-    }
+    public void sort(T[] array, int leftIndex, int rightIndex) {
+		int middleIndex;
 
-    private void merge(int[] arr, int lb, int mid, int ub) {
-        int i=lb, j=mid+1, k=lb;
-        int[] tempArr = new int[arr.length];
+		if (validation(array, leftIndex, rightIndex)){
 
-        while (i <= mid && j <= ub) {
-            if (arr[i] <= arr[j])
-                tempArr[k++] = arr[i++];
-            else
-                tempArr[k++] = arr[j++];
-        }
+			middleIndex = (leftIndex + rightIndex) / 2;
 
-        while (i <= mid)
-            tempArr[k++] = arr[i++];
-        
-        while (j <= ub)
-            tempArr[k++] = arr[j++];
+			this.sort(array, leftIndex, middleIndex);
+			this.sort(array, middleIndex + 1, rightIndex);
+			this.merge(array, leftIndex, middleIndex, rightIndex);
+		}
+	}
 
-        for (int l = lb; l <= ub; l++) {
-            arr[l] = tempArr[l];
-        }
-    }
-     private void printArray(int[] arr) {
-         for (int val: arr) {
-             System.out.print(val + " ");
-         }
-         System.out.println();
-     }
+	private boolean validation(T[] array, int leftIndex, int rightIndex) {
+		boolean isValid = true;
 
-    public static void main(String[] args) {
-        int data[] = {7, 9, 5, 2, -5, -1, -2, 0, 4, 3};
+		if (array == null || array.length <= 0) {
+			isValid = false;
+		} else if ((leftIndex >= rightIndex) || (leftIndex < 0) || (rightIndex <= 0)) {
+			isValid = false;
+		} else if ((rightIndex > array.length - 1) || leftIndex >= array.length) {
+			isValid = false;
+		}
 
-        MergeSort mergeSort = new MergeSort();
-        mergeSort.sort(data,0,data.length-1);
-        mergeSort.printArray(data);
-    }
+		return isValid;
+	}
+
+	private void merge(T[] array, int leftIndex, int middleIndex, int rightIndex){
+		int leftStart = leftIndex;
+		int leftEnd = middleIndex;
+		int rightStart = middleIndex + 1;
+		int rightEnd = rightIndex;
+		int index = leftIndex;
+
+		T[] temp = Arrays.copyOf(array, array.length);
+
+		while (leftStart <= leftEnd && rightStart <= rightEnd){
+			if (temp[leftStart].compareTo(temp[rightStart]) < 0){
+				array[index] = temp[leftStart];
+				leftStart++;
+			} else {
+				array[index] = temp[rightStart];
+				rightStart++;
+			}
+
+			index++;
+		}
+
+		if (leftStart <= leftEnd){
+			System.arraycopy(temp, leftStart, array, index, leftEnd - leftStart + 1);
+		} else if (rightStart <= rightEnd) {
+			System.arraycopy(temp, rightStart, array, index, rightEnd - rightStart + 1);
+		}
+
+	}
+
 }
